Матчмейкинг книг

+++ books_backend/src/main/java/com/psuti/books/service/OfferListService.java
@@ -4,9 +4,11 @@ import com.psuti.books.dto.OfferListDTO;
 import com.psuti.books.model.Category;
 import com.psuti.books.model.OfferList;
 import com.psuti.books.repository.*;
+import com.psuti.books.security.UserPrincipal;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -19,15 +21,22 @@ public class OfferListService {
     private UserRepository userRepository;
     private StatusRepository statusRepository;
 
-    public OfferList create(OfferListDTO dto) {
+    public OfferList create(OfferListDTO dto, UserPrincipal principal) {
+        List<Category> categoryFromDTO = new ArrayList<>();
+
+        for (Long id : dto.getCategories()) {
+            categoryFromDTO.add(categoryRepository.findById(id).orElse(null));
+        }
+
         return offerListRepository.save(OfferList.builder()
                 .bookLiterary(bookLiteraryRepository.findById(dto.getBookLiteraryId()).orElse(null))
-                .user(userRepository.findById(dto.getUserId()).orElse(null))
+                .user(userRepository.findByEmail(principal.getEmail()))
                 .isbn(dto.getIsbn())
                 .yearPublishing(dto.getYearPublishing())
                 .createAt(new Date())
                 .updateAt(new Date())
                 .status(statusRepository.findById(1L).orElse(null))
+                .categories(categoryFromDTO)
                 .build());
     }
 
@@ -38,21 +47,23 @@ public class OfferListService {
     public List<OfferList> getAll() {
         return offerListRepository.findAll();
     }
-    public OfferList update(OfferListDTO dto) {
-        return offerListRepository.save(OfferList.builder()
-                .id(dto.getId())
-                .bookLiterary(bookLiteraryRepository.findById(dto.getBookLiteraryId()).orElse(null))
-                .user(userRepository.findById(dto.getUserId()).orElse(null))
-                .isbn(dto.getIsbn())
-                .yearPublishing(dto.getYearPublishing())
-                .updateAt(new Date())
-                .status(statusRepository.findById(dto.getStatusId()).orElse(null))
-                .build());
+
+    public void delete(Long id, UserPrincipal principal) {
+        if (userRepository.findByEmail(principal.getEmail()) == offerListRepository.findById(id).get().getUser())
+            offerListRepository.deleteById(id);
     }
 
-    public void delete(Long id) {
-        offerListRepository.deleteById(id);
-    }
+//    public OfferList update(OfferListDTO dto) {
+//        return offerListRepository.save(OfferList.builder()
+//                .id(dto.getId())
+//                .bookLiterary(bookLiteraryRepository.findById(dto.getBookLiteraryId()).orElse(null))
+//                .user(userRepository.findById(dto.getUserId()).orElse(null))
+//                .isbn(dto.getIsbn())
+//                .yearPublishing(dto.getYearPublishing())
+//                .updateAt(new Date())
+//                .status(statusRepository.findById(dto.getStatusId()).orElse(null))
+//                .build());
+//    }
 
     public OfferList assignCategoryToOfferList(Long offerListId, Long categoryId) {
         List<Category> categoryList = null;
