+++ books_backend/src/main/java/com/psuti/books/service/UserAddressService.java
@@ -4,17 +4,20 @@ import com.psuti.books.dto.UserAddressDTO;
 import com.psuti.books.model.UserAddress;
 import com.psuti.books.repository.UserAddressRepository;
 import com.psuti.books.repository.UserRepository;
+import com.psuti.books.security.UserPrincipal;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @AllArgsConstructor
 public class UserAddressService {
     private UserAddressRepository userAddressRepository;
     private UserRepository userRepository;
-    public UserAddress create(UserAddressDTO dto) {
+    public UserAddress create(UserAddressDTO dto, UserPrincipal principal) {
         return userAddressRepository.save(UserAddress.builder()
-                .user(userRepository.findById(dto.getIdUser()).orElse(null))
+                .user(userRepository.findByEmail(principal.getEmail()))
                 .addrIndex(dto.getAddrIndex())
                 .addrCity(dto.getAddrCity())
                 .addrStreet(dto.getAddrStreet())
@@ -23,16 +26,20 @@ public class UserAddressService {
                 .AddrApart(dto.getAddrApart())
                 .isDefault(dto.isDefault())
                 .build());
+    }
+
+    public List<UserAddress> getByUserPrincipal(UserPrincipal principal) {
+        return userAddressRepository.findByUserId(userRepository.findByEmail(principal.getEmail()).getId());
     }
 
     public UserAddress getById(Long id) {
         return userAddressRepository.findById(id).orElse(null);
     }
 
-    public UserAddress update(UserAddressDTO dto) {
+    public UserAddress update(UserAddressDTO dto, UserPrincipal principal) {
         return userAddressRepository.save(UserAddress.builder()
                 .id(dto.getId())
-                .user(userRepository.findById(dto.getIdUser()).orElse(null))
+                .user(userRepository.findByEmail(principal.getEmail()))
                 .addrIndex(dto.getAddrIndex())
                 .addrCity(dto.getAddrCity())
                 .addrStreet(dto.getAddrStreet())
@@ -41,10 +48,10 @@ public class UserAddressService {
                 .AddrApart(dto.getAddrApart())
                 .isDefault(dto.isDefault())
                 .build());
-
     }
 
-    public void delete(Long id) {
-        userAddressRepository.deleteById(id);
+    public void delete(Long id, UserPrincipal principal) {
+        if (userRepository.findByEmail(principal.getEmail()) == userAddressRepository.findById(id).get().getUser())
+            userAddressRepository.deleteById(id);
     }
 }
