+++ books_backend/src/main/java/com/psuti/books/service/WishListService.java
@@ -1,15 +1,16 @@
 package com.psuti.books.service;
 
 import com.psuti.books.dto.WishListDTO;
+import com.psuti.books.model.Category;
 import com.psuti.books.model.WishList;
-import com.psuti.books.repository.StatusRepository;
-import com.psuti.books.repository.UserAddressRepository;
-import com.psuti.books.repository.UserRepository;
-import com.psuti.books.repository.WishListRepository;
+import com.psuti.books.repository.*;
+import com.psuti.books.security.UserPrincipal;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -18,29 +19,42 @@ public class WishListService {
     private UserRepository userRepository;
     private StatusRepository statusRepository;
     private UserAddressRepository userAddressRepository;
-    public WishList create(WishListDTO dto) {
+    private CategoryRepository categoryRepository;
+    public WishList create(WishListDTO dto, UserPrincipal principal) {
+        List<Category> categoryFromDTO = new ArrayList<>();
+
+        for (Long id : dto.getCategories()) {
+            categoryFromDTO.add(categoryRepository.findById(id).orElse(null));
+        }
         return wishListRepository.save(WishList.builder()
-                .user(userRepository.findById(dto.getUserId()).orElse(null))
+                .user(userRepository.findByEmail(principal.getEmail()))
                 .createAt(new Date())
-                .status(statusRepository.findById(dto.getStatusId()).orElse(null))
+                .updateAt(new Date())
+                .status(statusRepository.findById(1L).orElse(null))
                 .userAddress(userAddressRepository.findById(dto.getUserAddressId()).orElse(null))
+                .categories(categoryFromDTO)
                 .build());
+    }
+
+    public List<WishList> getAll() {
+        return wishListRepository.findAll();
     }
 
     public WishList getById(Long id) {
         return wishListRepository.findById(id).orElse(null);
     }
 
-    public WishList update(WishListDTO dto) {
-        return wishListRepository.save(WishList.builder()
-                .id(dto.getId())
-                .user(userRepository.findById(dto.getUserId()).orElse(null))
-                .status(statusRepository.findById(dto.getStatusId()).orElse(null))
-                .userAddress(userAddressRepository.findById(dto.getUserAddressId()).orElse(null))
-                .build());
-    }
+//    public WishList update(WishListDTO dto) {
+//        return wishListRepository.save(WishList.builder()
+//                .id(dto.getId())
+//                .user(userRepository.findById(dto.getUserId()).orElse(null))
+//                .status(statusRepository.findById(dto.getStatusId()).orElse(null))
+//                .userAddress(userAddressRepository.findById(dto.getUserAddressId()).orElse(null))
+//                .build());
+//    }
 
-    public void delete(Long id) {
-        wishListRepository.deleteById(id);
+    public void delete(Long id, UserPrincipal principal) {
+        if (userRepository.findByEmail(principal.getEmail()) == wishListRepository.findById(id).get().getUser())
+            wishListRepository.deleteById(id);
     }
 }
